backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
httpx==0.28.1
faiss-cpu==1.12.0
numpy==2.2.6
sentence-transformers==5.1.0
pypdf==5.9.0
requests==2.32.4
